feat: 예약 추가 및 삭제 기능 구현

+++ src/main/java/roomescape/dto/Reservation.java
@@ -3,5 +3,5 @@ package roomescape.dto;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public record Reservation(int id, String name, LocalDate date, LocalTime time) {
+public record Reservation(long id, String name, LocalDate date, LocalTime time) {
 }
